requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv

pydantic[email]

SQLAlchemy
asyncpg
psycopg2-binary

python-jose[cryptography]
passlib[bcrypt]

celery
redis

httpx
requests
pillow
